imglib2-core: Unsigned2BitType and Unsinged4BitType use a long index, and therefore can refer to up to Integer.MAX_SIZE * (64L / getBitsPerPixel()).
The question is what to do with methods int getIndex() and void updateIndex(int).
currently a cast to int is made.
These methods are never used for direct set/get of the index in the
underlying DataAccess in these 2 classes, and therefore could be using
long without issues, at least for ArrayImg derivatives.

+++ imglib2/core/src/main/java/net/imglib2/type/numeric/integer/Unsigned2BitType.java
@@ -55,7 +55,7 @@ import net.imglib2.util.Fraction;
 public class Unsigned2BitType extends AbstractIntegerType<Unsigned2BitType> implements NativeType<Unsigned2BitType>
 {
 	// Maximum count is Integer.MAX_VALUE * (64 / 2)
-	private int i = 0;
+	private long i = 0;
 
 	final protected NativeImg<Unsigned2BitType, ? extends LongAccess> img;
 
@@ -99,7 +99,7 @@ public class Unsigned2BitType extends AbstractIntegerType<Unsigned2BitType> impl
 	public NativeImg<Unsigned2BitType, ? extends LongAccess> createSuitableNativeImg( final NativeImgFactory<Unsigned2BitType> storageFactory, final long dim[] )
 	{
 		// create the container
-		final NativeImg<Unsigned2BitType, ? extends LongAccess> container = storageFactory.createLongInstance( dim, new Fraction( 2, 64 ) ); 
+		final NativeImg<Unsigned2BitType, ? extends LongAccess> container = storageFactory.createLongInstance( dim, new Fraction( 2, 64 ) );
 
 		// create a Type that is linked to the container
 		final Unsigned2BitType linkedType = new Unsigned2BitType( container );
@@ -122,7 +122,7 @@ public class Unsigned2BitType extends AbstractIntegerType<Unsigned2BitType> impl
 		return (dataAccess.getValue(k >>> 6) >>> (k % 64)) & mask;
 		*/
 		// Same as above minus one multiplication, plus one shift to multiply the reminder by 2
-		return (dataAccess.getValue(i >>> 5) >>> ((i % 32) << 1)) & mask;
+		return (dataAccess.getValue((int)(i >>> 5)) >>> ((i % 32) << 1)) & mask;
 	}
 
 	// Crops value to within mask
@@ -133,7 +133,7 @@ public class Unsigned2BitType extends AbstractIntegerType<Unsigned2BitType> impl
 		final long shift = k % 64;
 		*/
 		// Same as above, minus one multiplication, plus one shift to multiply the reminder by 2
-		final int i1 = i >>> 5; // k / (64 / 2)
+		final int i1 = (int)(i >>> 5); // k / (64 / 2)
 		final long shift = (i % 32) << 1;
 		// Clear the bits first, then and the masked value
 		dataAccess.setValue(i1, (dataAccess.getValue(i1) & ~(mask << shift)) | ((value & mask) << shift));
@@ -158,7 +158,7 @@ public class Unsigned2BitType extends AbstractIntegerType<Unsigned2BitType> impl
 	public double getMinValue()  { return 0; }
 
 	@Override
-	public int getIndex() { return i; }
+	public int getIndex() { return (int)i; }
 
 	@Override
 	public void updateIndex( final int index )
